Move components & pages folders outside app folder
