+++ app/layout.tsx
@@ -1,10 +1,10 @@
 import './globals.css'
-import Nav from './components/Nav'
-import Header from './components/Header'
-import Footer from './components/Footer'
+import Nav from '../components/Nav'
+import Header from '../components/Header'
+import Footer from '../components/Footer'
 import {getServerSession} from 'next-auth/next'
 import { authOptions } from '@/pages/api/auth/[...nextauth]'
-import Hydrate from './components/Hydrate'
+import Hydrate from '../components/Hydrate'
 import {Roboto, Lobster_Two, Dancing_Script, Castoro } from 'next/font/google'
 
 //Define main font
