+++ app/page.tsx
@@ -1,7 +1,7 @@
-import Discount from "./components/discount"
-import MainSection from "./components/main"
-import ProductSection from "./components/product-section"
-import DeliveryServices from "./components/DeliveryServices"
+import Discount from "../components/discount"
+import MainSection from "../components/main"
+import ProductSection from "../components/product-section"
+import DeliveryServices from "../components/DeliveryServices"
 
 export default function Main (){
   return (
