+++ components/product-section.tsx
